requirement.txt
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
PyJWT==2.7.0
SQLAlchemy==2.0.19